doc: Fix paper height
The original manual pages were 397 pt x 613 pt (5.51 in x 8.51 in),
which allows a book to be printed 2-up on US letter paper:

    $ wget https://cdn.cloudflare.steamstatic.com/steam/apps/450140/manuals/Seven_Kingdoms_Ancient_Adversaries_manual.pdf
    [...]
    $ pdfinfo Seven_Kingdoms_Ancient_Adversaries_manual.pdf | grep '^Page size:'
    Page size:      397 x 613 pts
    $ printf 'scale = 2; 397 / 72; 613 / 72;\n' | bc  # 72 pt / in
    5.51
    8.51

Note: In the original manual, the outside margin was larger than the
inside ("gutter") margin was.  But in book publishing, the gutter margin
should usually be larger, as it is in this manual, since some of the
gutter is lost to binding.  So leave the margins alone, because they're
more correct than the original manual's margins were.