Unit: Add hotkey to cancel build or settle
Patch originally from P. J. McDermott <pj@pehjota.net>

Modified patch to remove default key binding of Escape key. Escape
overlaps with a global key for dialogs, menus and reports, and you
should be using the right mouse button to cancel commands, as that
is what the game was designed to do.