News: Notify when firm/unit/weapon/ship is ready
It can be hard to hear when tasks are completed far away on the map or
to find a new unit or weapon among many villages or war factories.  This
change gives a more prominent visual notification with a button to go
and select the new firm or unit.

A similar idea using the mini map was proposed by "SheeEttin" in 2010:
<https://www.7kfans.com/forums/viewtopic.php?t=164>.  The mini map would
be hard to notice and understand, so this change uses the news instead.

News message formats:
* "Your <Firm> has been built."
* "Your <Skill> has been trained in <Town>."
* "Your <Weapon> <Tech Level Number> has been built near <Town>."
* "Your <Ship> has been built."

Units are trained in towns, and weapons are built in war factories
linked to towns.  But new firms may be constructed far from any town,
and ships may be built in harbors far from any town.  So only the unit
and weapon news messages reference towns.