NationAI: Don't duplicate build fort orders
This is found when the ordered unit has a long path to take to start the
build, the build order gets reissued because the same conditions applied
when think_capture_independent executes once again. The test case I had
the order was duplicated 4 times over the original.

7k2 has a very similar fix, and I opted to use the more generalized check
of the build order, as that seems like it will be more reliable in the
long run.