Unit: Downgrade stop order on handle_blocked_by_idle_unit failure.
Since commit 25b4b6fd4f4d0e136313fa87bb1e0d3d065a743c that added the fix
called fix_path_blocked_by_team, handle_blocked_by_idle_unit is now used.
Calling this function can trigger an assertion at
Unit::on_way_to_attack OUNITAT2.cpp:873:
	err_when(action_mode!=ACTION_ATTACK_UNIT ||
		action_para!=targetUnit->sprite_recno);

The assertion is triggered because the unit is able to path, but the path
goes through an idle friendly unit, and then the stop order seen is
issued. This stop order clears the attack action, which is not expected
per the original developer.

It would seem that it makes no sense to expect the order can continue if
the unit cannot fulfill it being blocked. However, in review of the
action code, pathing issues are seen usually to not clear out the action
as the stop(KEEP_DEFENSE_MODE) does. In addition, it would make sense to
give the unit the ability to recover at some point.

So instead of removing the assertion, a better option seems to be to use
stop(KEEP_PRESERVE_ACTION) because that does not clear out the action. In
fact, this is usually what is done because it simply just turns the unit
back to idle, but the action is still available to reactivate.

Usually stop(KEEP_DEFENSE_MODE) or stop(0) is used in cases where the unit
is issued a completely new order from player action, or the order is
completely impossible and not temporary.

This rolls a change here into the same fix of the mentioned commit.