NationAI: Fix targeting attacker when it hides in a building.
One test case showed a fleeing general who lost its fort temporarily
stopping to fire a shot at an enemy firm. It just so happened to be also
next to another empty fort of its own nation, and then decided it needed
to hide in that fort, and left the map. The bullet strikes the fort, and
the code is unable to properly track down the attacker.

But because the same square was also immediately occupied by an enemy
soldier, an assertion that I inserted a few months ago to make sure that
the defend code does not try to set the at war flag on itself identified
there was an issue. The soldier now occupying the square of the attacker
was of the same nation of the fort that struck. In a worse scenario, this
could have lead to a war status with the wrong nation, and the assertion
would not have caught it.

There is not an easy fix here, because the unit code does not have a unit
mode (i.e. UNIT_MODE_WORKER similar to UNIT_MODE_CONSTRUCT) to identify
where the unit really is, the location has to be searched for somehow. So
for now the fix will be to search around the point of the last known
location for the firm that the unit hid in and order to attack that. The
code is not specific to the type of unit, but that is only because the
better fix would not need additional code for that anyway.

7k2 has the same problem.