Town: Canceling training when a town turns independent.
Independent towns can't train units. But they can if it was in a prior
nation and turns independent. This was a surprising common issue, but not
found until in one case a spy was in training and hit a corner case in
the spy ai code when the spy was orphaned through the transition.

We'll keep the designer intent that training can continue, but now not
when a town turns independent.

7k2 does not have the same bug.