NationAI: Improve detection of targeted nation.
A case was found during randomized testing with an ai using a greater
being to attack, and a defender appearing under it. The defending ai tried
to target itself due to having the defender underneath instead of the ai
with the greater being. This change updates the
enable_should_attack_on_target and get_target_nation_recno ai functions
due to the following reason.

* The ai is attacking an air unit, and end up targeting a land unit
  instead.
* The ai may accidentally try to attack itself instead of the air unit
  overhead.
* The ai is unable to properly attribute the target nation if a land unit.
* The ai is unable to properly attribute the target nation if a sea unit.
* The ai might later not clear the attack if originally targeting an
  independent unit.

Fixes:
* Return -1 for no target nation. 7k2 has similar fix.
* Do not return a nation_recno if the object nation_recno is the ai nation
  itself. 7k2 does not have this fix.
* Acknowledge AIR and SEA units present as an option. AIR units have
  priority over LAND. 7k2 has a partial fix like this.