Fix memory leak on loading (Nation::action_array)
When loading, action_array in the Nation class, which is a DynArray,
would leak memory.

This was due to action_array (as a DynArray) allocating memory in its
constructor, and then subsequently having the DynArray overwritten by
all zeroes or the loaded data from a save file version 1 nation, losing
the pointer to the memory that was initialliy allocated.

Fix this by freeing the memory right before we zero out the action_array
DynArray or load a version 1 nation.