Nation: Split out file records for GameFile.
Items split out are NationArray, Nation, and dependent classes and structs.

I/O is layered on the original File class, and the process of reading and
writing mirrors the original GameFile code, with new code that doesn't
overwrite the class instance data members directly. The record is packed as
the original GameFile record, and unpacked into the Nation class adhearing
to differences in the structure format, and not destroying important data
that needs to be left alone.

Byte swapping is implemented for class primitive data members and complex
data members as far as it was part of the record in/out. Some items were
skipped in the original code and handled separately, like dynamic arrays.
So not all parts of the class are finished just yet.

More classes to come.

This also fixes two places that action_array memory was not freed.

A few notes on the work in progress:
Byte swapping is not implemented for the File class put/get and simple
arrays of put/get of char,short,long. It will be done later.

Byte swapping is not done for Crc computation. It will be done later.

Additional file formats, such as replay files, will need review.

What is primarily being addressed now are all calls to file_read() and
file_write(), as these are not portable unless we pack and unpack with
byte-swapping first, hence defining the records and doing just that.