Set config flag when incompatible with 7kaa
This allows 7kaa to recognise when connecting to a 7K: Ambition
instance if it is not running in a 7kaa compatibility mode.

This will need to be made more robust later on to update the flag when
the mode changes, as the 7kaa flag gets set only once on
initialisation.