Fix unit path line first segment going backwards
The first segment of units' path lines would appear to flow backwards,
while the rest of it would flow forwards.

This was because the start of the line (the unit) is continuously moving
and the rate of movement makes the path line appear to flow backwards.

Fix this by making the phase depend on just the coordinates where the
line is being drawn instead, meaning that it is always fixed, regardless
of the start and end points.

I suspect that this is why the original (Enlight) code chose to draw the
line back-to-front instead of flowing the correct way.

Another approach to consider is just drawing from the tail-end instead.