Highlight selected units
To make it easier to tell which units are selected, highlight all of the
currently selected units, in a similar manner to buildings.

Concurrently, as we are using the cyan colour for selected units, change
the hover colour to yellow.  As it happens, this better fits with the
theme from the main menu and elsewhere.