Fix scrolling right and down
Scrolling right and down only worked as long as the user kept moving the
mouse towards the right or down, as opposed to left or up, where keeping
the cursor against the edge of the screen was sufficient.

This was due to a weird interaction between SDL scaling and the defined
mouse boundary.

Fix this by making scroll area one pixel bigger inside to catch the
cursor even if its position gets rounded out and pushed in by a pixel.