Fix multiplayer load game waiting for user input
The multiplayer load game setup screen would be unresponsive and not
update players' ready status or messages and not start the game until
there was user input when running in enhanced mode or higher.

This was due to the delayFrame call having no timeout, and so waiting
indefinitely for user input.

Fix this by giving this screen the same timeout as the other multiplayer
screens.