GameFile: Restore as object and split out file record.
In order to implement the packing and unpacking in the expected manner,
the class really should be treated as an object, which is how it was
originally. Most of the passing around the game file header as a
completely independent struct can just be cleaned up. There will be more
work to reduce the code handling game files.