Fix in-game menu not showing straight away
The in-game menu would not show straight away after clicking it.

This was because the loop for the in-game menu actually renders to the
screen before drawing the menu itself, meaning that it did not display
on the first cycle before delaying.

Fix this by just running another cycle (two in total) before delaying.