Fix in-game menu options not showing straight away
The option, save and load game, and tutor options did not show straight
away after clicking them.

For the option screen, this was because the loop for it renders to the
screen before drawing the menu itself, meaning that it did not display
on the first cycle before delaying.  For the rest, they were missing the
first-cycle check.

Fix this by running another cycle (two in total) before delaying for the
option screen, and delaying by one cycle for the other screens.