GameFile: Update town and firm links on load.
With the distance calculation change for town and firms (48c37f1) in save
files that were produced in the old game versions, they could actually be
in reach now. This leads to a situation like in the scenario "A Weakened
State" an ai war factory now links to another town, and migration is now
considered where it was not possible before. Links are formally created
when the place is created or destroyed. So it led to a situation where
workers migrate to a town without a formal link, throwing off the pop
stats when checking the list of linked firms, and triggering an error.

Update the links at the game load to have proper accounting present when
linked firms are processed.