Update Firm::last_attacked_date on all damage
Firms' `last_attacked_date`s were not being updated on all damage
because there are various checks to prevent unnecessary action when a
shooting unit already died, a porcupine exploded, or a unit changed
nation.

However, this meant that the various checks which rely on
`last_attacked_date` did not realise that the Firm was, in fact, taking
damage.

By updating the value on all damage, any checks that rely on
`last_attacked_date` will now be correct.