Fix own spy icon not showing all the time
The spy icon used to show all the time, regardless of whether the unit
was selected or the icon showing option setting, and this was broken in
a previous change.

This was due to overzealous trimming and insufficient testing.

This reverts commit 077f16e3450fb6126032aea9faecf4828a2e2479 "Remove
unnecessary code for drawing spy icons" and adds a line of code.