Fix line animation on game speed unlimited (9)
The line animation speed was breaking on game speed unlimited (9).  Some
users also experienced this issue when moving the mouse.

This was due to 7kaa's disp_frame() being called even when a video frame
is not to be displayed.

Fix this by making the phase depend only on elapsed time, as with other
animations.