Fix Inn hire button being en/disabled incorrectly
The hire button in the inn would sometimes be disabled or enabled even
when the selected unit could be afforded or not, respectively.

This was due to the logic for enabling and disabling the button not
being updated to use the inverse as with the rest of the inn code.