Fix info pane hitbar pixel corruption
The hitbars in the info pane would get corrupted and show extra pixels
to the right of them when a unit entered or left the building.

This was due to not drawing anything new to the buffer, and so the
previously drawn hitbar sitting there and being slightly drawn over.

Fix this by blitting from the back buffer onto the front for the
remaining unused width.