Fix git_string.h missing on first build
When building fresh for the first time, it would fail with an error
about git_string.h missing:

    Ambition_version.cc:28:10: fatal error: git_string.h: No such file or directory
       28 | #include "git_string.h"
          |          ^~~~~~~~~~~~~~
    compilation terminated.

This is probably because of the order that make runs the files in, as
deleting git_string.h later does not cause this issue.

Fix this by explicitly telling make that git_string.h a dependency of
Ambition_version.cc, so that it's always generated.