Fix requested builder unable to reach destination
When the nearest builder Unit, getting requested to travel to a Firm,
was inside a Firm which was in a different region to the destination
Firm (meaning that it was not reachable by walking), the builder would
start the journey and then get stuck along the way, resulting in no
builder in the requested Firm and the builder standing around idle and
losing loyalty.

This was due to the region ID being checked only for builders outside
Firms.

Fix this by checking the region IDs of the Firms the builder Units are
inside if they are inside one.