Fix technologies researching beyond level 3
Technologies were researching beyond level 3 when Towers of Science were
left alone.

This was due to there being no check for Towers of Science about whether
the research they are undertaking is appropriate.

Fix this by checking whether the research is possible before sending it
to the Tower of Science.