Fix multiplayer desync when using unit waypoints
Multiplayer would desync when a unit used waypoints.

This is due to move commands not being processed immediately when they
are issued in multiplayer, meaning that the rest of process_idle() would
run in 7kaa, including various frame counts.  Although this was actually
inconsequential, it would cause a momentary desync.

Fix this by running the rest of process_idle() before ordering the unit
to the next waypoint.