Work around waypoints disappearing in multiplayer
In multiplayer, when a unit had a series of waypoints, the first ones
would disappear and the unit would follow a later one.

This was due to the delay in multiplayer between issuing an order and it
actually taking effect, meaning that the waypoint had been popped, but
the unit was still treated as idle.

Work around this by adding a delay between popping waypoints.