Fix crash when a kingdom is destroyed or surrenders
When a kingdom was destroyed or surrendered while still having spies,
the game would crash.  This crash also happened on exit.

This was due to untracked spies getting dropSpyIdentity called, which
resulted in segfaults, since findBy7kaa* was correctly returning a null
pointer when it did not find a tracked Ambition::Unit.