World: Fix scrolling by allowing larger scroll region due to warp slop.
Tests showed that mode switching has bad behavior in current versions of
Windows. So we need another way to accommodate people on their various
screen resolutions. Added a setting to increase the region where mouse
edge scrolling will occur. This is needed since we use mouse warping to
keep the mouse within a certain region, and due to window stretching and
the virtual coordinate system, the warp function is not precise enough for
the scroll detection exactly on the window edge.

Perhaps SDL3 will fix the imprecise calculations, but still, we should
just revamp the code to support flexible screen sizing, then the virtual
coordinate system is not needed.