Fix not being able to attack porcupines
Shift-right-clicking on a porcupine added a waypoint instead of
attacking it.

This was due to the change allowing shift to be used to create waypoints
causing Power::detect_action() to return early.

Fix this by checking whether the target is a porcupine in
Ambition::Unit::toggleWaypoint() and running the 7kaa waypoint code only
on ALT, not shift.