Fix production queue count with 7kaa queue
When there was a 7kaa queue (i.e. loading a 7kaa game with a queue
already in place), the queue count would show as zero instead of the
actual amount.

This was due to the 7kaa queue count being disregarded in the code,
which is correct only on the assumption that all the queuing is done
through Ambition—but this is not necessarily the case when loading an
existing save game.

Fix this by simply adding to the 7kaa queue count instead of
disregarding it.