Fix inability to cancel 7kaa queue
After loading a save game with an existing 7kaa production queue, right
clicking to reduce the size of the queue would cancel only the current
production and have no effect on the queue itself.

This was due all queue manipulation code only working with the Ambition
code and ignoring the 7kaa queue.

This is correct behaviour, and we do not want to be mixing the queues,
so fix this by migrating the 7kaa queue to an Ambition queue whenever we
need to actually manipulate the queue.