Fix production queue count showing zero with 7kaa queue
When there was a 7kaa queue (i.e. loading a 7kaa game with a queue
already in place) but no Ambition queue yet, the queue count would show
as zero instead of the actual amount.

This was due to the 7kaa queue count being disregarded in the code.

Fix this by simply returning to the 7kaa queue count instead of zero.