Fix pixel artefact bars at the bottom of pictures
The bottom of the credits and results screens had coloured (depending on
the colour palette) bar artefacts.

This was due to an off-by-one error in VgaBuf::put_large_bitmap(), which
caused less than the entire file to be read in, resulting in missing
pixels at the end.  This was previously masked by (probably) no images
being small enough to fit into the buffer and follow this code path.