Work around spy info pane corruption, including change nation
When there are two rows of colours for spies, the info pane was getting
corrupted, with the lower row of buttons getting too low and leaving
artefacts.  In addition, the clicking of the change nation colour
buttons was not registering correctly.

This was due to the change nation panel taking up too much space now
with the leadership panel as well.  The clicking of change nation
buttons was not working because the drawing and click detection
co-ordinates are not connected.

Work around this for now by just not showing the leadership panel for
spies.