Fix duplicated spy icon
When workers were re-arranged inside firms, whether by skill change or
workers entering or exiting the building, the spy icon could get
duplicated if the spy moved and the worker taking his place had the same
race.

This was due to an optimisation that was putting in now not being
possible due to the spy icon being placed over the portrait.

Fix this by bypassing this optimisation.  The savings are minimal on
today's hardware.