Camp: Fix high-potential soldiers gaining less leadership
When a soldier had a high leadership potential, his leadership skill was
actually climbing slower than a soldier with lower leadership potential.

This was due to an overflow in Worker::skill_level_minor, which is a
signed char, and was overflowing past 127.  As high-potential soldiers
could gain even up to 180 minor levels each training cycle, most of the
time the overflow would mean that it took them far longer to gain
levels.

Fix this by using the same levelMinor and while loop as the combat
level, meaning that only a value less than 100 would actually be stored
in skill_level_minor.