Firm: Fix units losing their skill potentials
When entering a building, unit skill potential was getting changed from
what it was before they entered.  And when leaving a building, the
potential was reset to zero.

This was due to the potential mistakenly getting re-rolled when a unit
entered a building, meaning that any units that had potential were
likely to lose it, and those without it had a chance of gaining it.  At
any rate, the potential was almost certain to be different than before
the unit entered the building.

On the converse, when leaving the building, the potential was not
getting set at all.

[Added note: 7k2 rewrote skill code, likely fixing the issue there]