Fix back buffer artefacts from disabled buttons
The back buffer was having artefacts left over on it from disabled
buttons.

This is due to the back buffer being used temporarily to darken buttons
before copying them onto the front buffer.  Before this, the area where
the button will be is saved into a buffer and then copied back, to
restore the back buffer to its previous state.  However, this area was
slightly too small (52×43 pixels) compared to the button bitmap
size (51×46 pixels), meaning that three lines were left behind as an
artefact on the back buffer.

Fix this by increasing the size of the saved area, to fully cover the
area where the bitmap gets put.