Fix crash on research target after opening military report
Building a Tower of Science would cause a crash if the military report
were opened before setting a global research target.

This was due to Polity::_researchTarget being uninitialised and there
being no check on the value of the research target.  Even when it
happened to be set to 0, accessing tech_res with an index of 0 is
actually still a memory violation.

Fix this by initialising _researchTarget properly and also checking
researchTarget()'s return value before using it to index tech_res.