Fix Town training icon not appearing in info pane
The training icon (and the panels) were not appearing in the info pane
for Towns unless the Town were clicked on again.  And then, after one
unit finished training, the icon would disappear again.

This was due to Ambition using Town::recruit() directly (to get the free
remote handling) instead of the Town::add_queue() that the buttons use
in 7kaa, which calls Town::process_queue() and eventually info.disp().

This was only a bug in single-player, not multi-player, as the remote
message for MSG_TOWN_RECRUIT actually calls add_queue() instead of
recruit().

Fix this by calling info.disp() directly after calling recruit() in
Building::produce().