Fix memory violation with building guides
When placing a building near the bottom or right of the world, memory
violations would occur.  This did not affect users as it was only reads,
not writes, and any false positives would still be countered by the
border tiles being unbuildable.

This was due to two factors:

1. The world bounds being passed in were slightly too large, due to an
off-by-one error.

2. The inner loop checking build points not actually checking bounds.