Fix Town training menu (and others) exiting when training
The Town menu mode kept getting reset to the main menu whenever a unit
was trained, preventing the user from doing whatever they were doing.

This was due to Town::disp_info() for some reason resetting the menu
mode back to main if the refresh mode is set to REPAINT.

In other calls to disp_info, this is worked around by setting
disable_refresh.

Fix this by calling disp_info() only if the menu mode is already set to
main, the same way that process_queue() does.