Fix building guide lines not reaching viewport edges
The building guide lines were not reaching the top and right edges of
the viewport by one pixel.

This was due to an off-by-one error in the calculation of the viewport
width and height.

The effect on the user is small and possibly not noticable for more, but
more importantly, this would affect all future work that depends on the
viewport.