Fix firm frame out of bounds index
calculateFirmFrame() was causing an out of bounds index access.  There
seems to be no noticable impact on graphics.

This was due to passing a raw index to Firm::frame_delay(), which
actually subtracts one from the value passed, resulting in an access of
-1.

Fix this by just adding 1 to the index before passing it to
frame_delay().