Fix Inn scrolling when there are more than 4 guests
When there are more than 4 guests, scrolling would select the lower
guests, but would not scroll the browser properly, so would appear off
the bottom.  The same was true for scrolling up, if first moving the
browser down.

This was due to top_rec_no not being updated for the Inn guest browser.
Which was not caught due to insufficient testing.