Add Linux packaging script.
This follows the same basic strategy as the Windows packaging script. There
are some common libraries, most shared, and enet is linked statically. It
is probably a good idea to try to ship the same version enet in each
release across platforms.

The user will need to provide their own copies of the common libraries,
since distros usually will install them for the user. We will want to see
how this goes.

The FHS paths can be disabled, and so any install of the package won't
accidentally pick up any distro versions files that might be different.

There is no installer yet. But it can be added.

The executable will pick up the game data automatically when placed just
below the game data, much like originally with the CD. This allows us to
just place a symbolic link to the game if the user wants it in their path,
or other ways. I'd like this to expand to all the platforms and distro's
packaging actually. But for another time.