Fix Town training button artefacts and size
The skill training buttons on the Town training menu were too large and
going too far down the info pane, leaving behind artefacts after leaving
the menu.

This was due to BUTTON_ACTION_HEIGHT being adjusted to the correct value
for action buttons; but it seems that not all of its uses were checked.
One place where it was incorrectly used was for the training buttons,
which have now increased in size.

Fix this by correctly using a separate constant—TRAIN_BUTTON_HEIGHT—for
the training button heights.