Fix date and time display for old save files
The date and time for save files from 7kaa 2.15.6 (and apparently all
the way back to 2.14.x) were showing as "1 Jan 1601 09:39", regardless
of when they were actually saved.

This was due a change made in 7kaa 2.15.7 where the file time is written
into and read from inside the file itself instead of the file's system
time, which was written as 0 from 7kaa 2.14.x.  More information can be
found in [this forum thread](0).

Fix this by reading the file's system last write time when the 7K header
data file time is set to 0 (i.e. the old versions).

[0]: https://7kfans.com/forums/viewtopic.php?t=1571&sid=f6a6f3bd3dae1f54ff44d272d8ffc534