Fix spies always getting caught after stealing reports
Regardless of the spy skill of the spy, the first three steal report
attempts after loading the game would always result in the spy getting
caught.

This was due to the random number generator (misc2) always being
initialised to 0 when the game starts and never being updated except
when getting a random number for the steal report escape chance.  Thus,
the first three results would all be 0, which, because of the escape
chance algorithm, meant that the spies always got caught.

Fix this by initialising misc2's random number generator on start using
the current time.