Fix save file name and date artefacts off the right
If a file name were longer than usual, or the language set to one where
the file name or date were longer than in English, the name and date
would run off the right side of the load screen and create an artefact,
making it illegible.

This was due to just printing the text from a fixed point towards the
right without taking into account the text length.  There was some sort
of mitigation for this with FRENCH and GERMAN #defines, but they do not
work with the dynamic language switching that we currently have.

Fix this by properly printing the file name and date as right-aligned
text.