Fix crash on selling an empty Camp with a rally point
Selling or demolishing an empty Camp which has a rally point would cause
a segfault crash.

This was due to the array of units which are sortied from the
Camp (which is empty) being indexed outside its bounds.

Fix this by checking the emptiness of the array before it gets used.