Sys: Fix steal report escape not depending on report
The chance for a spy to escape after stealing a report was too high for
some spies and too low for others.

This was due to the incorrect variable being passed to
spy_array.needed_view_secret_skill(), meaning that the Spy's record
number was used for the report type, meaning that spies 1–5 had a pretty
hard time, spy 6 would always get caught, and the other ones had an
extremely high (83%) chance of escape.

Fix this by passing the correct variable to
spy_array.needed_view_secret_skill().