Fix build guidelines fading in colour on highlighting
The right and bottom build guidelines were appearing faded as compared
to the other guidelines when they were over terrain highlighting.

This was due to the way the human eye perceives colours relative to each
other and because these guidelines were in between pixels of
highlighting instead of over them, like the other guidelines.

Fix this by drawing these guidelines to the left and up, over the top of
pixelisation, instead of in between.

This has resulted in the guidelines being shifted over one pixel, which
can be noticable at times.