Work around Inn ghost guest crash
Rarely (on some machines), Inns ended up with a ghost guest that would
usually cause a segfault.  Though, sometimes, it would just show as an
empty record.

This was due to the `total_rec_num` of the Inn's `browse_hire` getting
out of sync with `inn_unit_count`, which would result in
getInnSelectedRecordNumber() returning 0, and then a negative indexation
(well, actually, pointer arithmetic) of `inn_unit_array` in
put_hire_rec().

The cause of this desync is still not clear.

Work around this in the meantime by adding bounds checking into
getInnSelectedRecordNumber().