Fix save game list scrolling further than expected
The save game list always seemed to have a delay on scrolling where it
would scroll one more than expected, resulting in the list jumping when
trying to select a specific slot.

This was due to the frame buffers not actually being flipped after the
frame was drawn and the frame being held in the back buffer until
something else (such as moving the cursor) caused an update.

Fix this by putting the vga.flip() call into the correct place, after
drawing the frame.