Work around terrain not matching its pathing
A number of Dirt types looked like they should block building and
movement but did not, and vice versa.

This may have been an oversight when they were previously put in.  It is
not clear what else could have caused this.

The ideal solution to this would be to change these Dirt types to either
blocking or non-blocking, to match their visual appearance.  However, as
this would break compatibility with 7kaa and as the 7kaa devs do not
want to fix the blocking types of the Dirt types, we need to work around
this by changing the appearance of the Dirt to match their blocking
types.