Make the corners and edges of Dirt terrain look impassable
The corners and edges of a lot of Dirt types looked like they were
passable, even thought the entire rectangle is not buildable or
walkable.

This was just part of the original design.

The ideal solution to this would be to change the shapes of the Dirt
types to match their visual appearance.  However, as this would break
compatibility with 7kaa and as the 7kaa devs do not want to fix the
blocking types of the Dirt types, we need to work around this by
changing the appearance of the Dirt to match their blocking types.  This
is done by adding squares from other various Dirt types (including the
now-unused ones that were changed in the fix for the complete
passability/impassibility.