Fix cloning caravans appearing to not work
Clicking on the clone caravan button in the info pane or on the trade
report would sometimes appear to do nothing and the caravan would not be
cloned.

This was actually correct behaviour, because idle caravans are used in
the cloning process, and attempting to clone an idle caravan just copies
its route to another (or the same) idle caravan.

Fix this by disabling/hiding the clone caravan button for idle caravans.