Fix Fryhtan Lair off-centre drawing
Fryhtan Lairs, especially Ick and Wyrm Lairs, looked very off-centre
relative to the space they actually take up, and when surrounded by
units.

It is unclear why this was the case.  The bitmap information has offset
values which should have been used to adjust them to the correct
locations.

Fix this by offsetting the base locations that the bitmaps are drawn.