Fix build Seat of Power keybind not working
When trying to press B-P to build a Seat of Power, the game was instead
toggling opaque report mode.

This is because the keybind for building a Seat of Power (P) was
conflicting with the toggle opaque report mode keybind.

Fix this by clearing the default keybind for toggling the opaque report
mode.  Users can set their own keybind in the config.txt file if they
wish.