Fix assassination estimate being too low on spy skill
As it turns out, the assassination success estimate was too low based on
the spy's skill.

This is because in 7kaa, the spy's skill is actually counted twice—once
in the beginning with a factor of 1, and once again when looping over
all of the Firm Workers with a factor of 0.25.