Fix production queue size display delay
The production queue size displayed on buttons seemed to have a delay in
its update, where it seemed to update with the old number when a unit
finished production.

This was due to there being an info.disp() run after the unit training
was complete, but before another unit from the queue was put into
production.

Fix this by running info.disp() after the queue is popped.