Fix crash after unloading scenario twice
After playing and exiting a scenario, playing and exiting a scenario
again resulted in a crash due to a double-free.  Also, attempting to
connect to the multiplayer service would fail.

This was due to config_adv.mp_service_addr being freed after exiting a
scenario, because the destructor of a copy of ConfigAdv is run, which
was pointing to the same address as the main ConfigAdv, but the address
continues to be used in the main config_adv, resulting either in the
double-free crash when exiting a scenario, or reading garbage when
attempting to connect to the multiplayer service.

Fix this by setting the mp_service_addr of the copy of ConfigAdv to NULL
before it goes out of scope and its destructor is run, preventing it
from freeing the memory.