Fix production queue counts bouncing in multiplayer
The production queue count displays would bounce down and up in
multiplayer when a new unit started production.

This was due to the latency in 7kaa's multiplayer handling between an
action being input and taking effect, which resulted in a delay between
the production queue being popped and the production actually starting,
during which time only the production queue size without the current
production was being taken into account for display.

Fix this by holding the queue and not actually popping it until we have
confirmation that the production has actually started or been put into
the 7kaa queue.