Fix own kingdom not selected for score comparison
The score comparison in the ranking report was not starting with the
player kingdom selected on the first run.  And would select other
kingdoms if used in multiplayer.

This was due to the game state reset happening after the game run,
rather than before, and therefore the selected comparison reset
happening at the wrong time.

Fix this by resetting the game state before the main loop of the game
instead of after.