Fix desync when built with gcc 14
Binaries built with gcc 14 were desyncing with binaries built with gcc
12 and the 7kfans releases (which are apparently built with gcc 11.2)
resulted in desyncs in the town_array.

This was because the default for `-fexcess-precision` changed in gcc 14
from `fast` to `standard`.  This means that floating point operations
were subtly different because of rounding happening at slightly
different times.

Fix this by explicitly setting `-fexcess-precision` to `fast` to always
have the same behaviour with gcc 14 or earlier.