Fix crash on ship list on Trade Report
When having more caravans than trade ships (which is almost always the
case), opening the ship list on the Trade Report would cause a segfault.

This was due Trade::drawReportCaravanCloneButton() running in both
modes, but the Browser holding either caravan or ship record numbers;
and when there are usually fewer ships than caravans,
info.get_report_data() actually overruns.

Fix this by not running Trade::drawReportCaravanCloneButton() when the
Trade Report is not in caravan mode.